Separated the console code from the formatting code
