+++ src/ConsoleColorOutput/OutputTextProcessor/IOutput.cs
@@ -14,15 +14,43 @@
 		/// </summary>
 		int Height { get; }
 
+		void StartParagraph();
+
+		void EndParagraph();
+
+		void SetColor(OutputColor color);
+		
 		/// <summary>
 		/// Output a block of text
 		/// </summary>
 		/// <param name="str"></param>
-		void Output(string str);
+		void WriteText(string str);
+
+		void WriteEmptyLine();
 		
 		/// <summary>
 		/// Pause output
 		/// </summary>
 		void PauseOutput();
+	}
+
+	public enum OutputColor {
+		Default = -1,
+		
+		Black = 0,
+		
+		Gray = 7,	
+		
+		Blue = 9,
+		
+		Green = 10,
+
+		Cyan = 11,
+		
+		Red = 12,
+		
+		Yellow = 14,
+		
+		White = 15,
 	}
 }
